Complete Admin Implementation
b. Create a new model called ‘role’ and create an additional model
called ‘user_role.’  The ‘user_role’ model will be the join table. Use
a many to many association so that a user has many user_roles, and a
role has many user_roles. The user_role model should belong to both
user and role.
c. Add a an attribute of admin(string) to your ‘role’ model.
d. Create a method to check if a user is an admin in your ‘user’ model.
You can use a method called ‘def admin?’